Fix termination issue with `wordlist` argument.
The application did not terminate upon consuming the entire `wordlist` due to the behavior of `read_line()` at EOF (it returns 0). This was resolved by enhancing the match statement in the `next()` method of the `LineProducer`. As a consequence, the return type of the `next()` function in the `Producer` trait was adjusted.