Fix: Trim newline char from lines read from file input.
The application did not remove the newline character returned by the read_line function. This led to incorrect password comparisons when using the `wordlist` argument.

Note: If the password legitimately ends with a newline character (which should be rare) and it is the last line in the file and the terminating newline, then this change could potentially remove a valid newline character.