Re #10 Create new search type that uses letters and words
Hitherto, searches were limited to numeric searches or dictionary
searches. The herein added search does a traditional brute-force search
using a currently-hardcoded character-set. The characters in the
characterset are a-zA-Z0-9, so passwords like EatMy59Pandas will be
covered. The search strategy takes a maximum length and uses four
characters as a minimum length.